backend/requirements.txt
torch
torchvision
Pillow>=8.0.0
